pyproject.toml
[project]
name = "r2copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "github-copilot-sdk>=0.1.23",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "streamlit[starlette]>=1.54.0",
]
